Move data generation functions to separate module
